+++ tests/single.test.ts
@@ -1,5 +1,12 @@
 import { test } from 'node:test';
 import { runOasTest } from '../src/tests/runners.js';
+import { OasGen } from '../src/oas/oasGen.js';
+import fs from 'fs';
+import _ from 'lodash';
+import assert from 'assert';
+import diff from 'deep-diff';
+// import { stringify } from 'flatted';
+import { stringify } from 'superjson'
 
 /*test('test_054_oas_test-better-naming', async () => {
 
@@ -11,80 +18,4 @@ import { runOasTest } from '../src/tests/runners.js';
 
   await runOasTest('launch_Library_2-docs-v2.3.0.json', paths, 116, 5);
 });
-
-test('test_019_oas_test_010_TMF633_IntentOrValue_to_Union', async () => {
-  const paths = [
-    'get:/product/{id}>res:r>comp:type:#/c/s/Product>comp:type:#/c/s/Entity>obj:type:#/c/s/Addressable>prop:scalar:id',
-    'get:/product/{id}>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:comp:intent>union:#/c/s/IntentRefOrValue>comp:type:#/c/s/IntentRef>comp:type:#/c/s/EntityRef>obj:type:[inline:#/c/s/EntityRef]>prop:scalar:@referredType',
-    'get:/product/{id}>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:comp:intent>union:#/c/s/IntentRefOrValue>comp:type:#/c/s/IntentRef>comp:type:#/c/s/EntityRef>obj:type:[inline:#/c/s/EntityRef]>prop:scalar:id',
-    'get:/product/{id}>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:comp:intent>union:#/c/s/IntentRefOrValue>comp:type:#/c/s/Intent>obj:type:[inline:#/c/s/Intent]>prop:scalar:name',
-    'get:/product/{id}>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:comp:intent>union:#/c/s/IntentRefOrValue>comp:type:#/c/s/Intent>comp:type:#/c/s/Entity>obj:type:#/c/s/Addressable>prop:scalar:id',
-    'get:/product/{id}>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:comp:intent>union:#/c/s/IntentRefOrValue>comp:type:#/c/s/Intent>obj:type:[inline:#/c/s/Intent]>prop:scalar:description',
-  ];
-  await runOasTest('TMF637-001-UnionTest.yaml', paths, 1, 4);
-});
-
-test('test_020_oas_test_010_TMF633_IntentOrValue_to_Union_Full', async () => {
-  const paths = ['get:/product/{id}>**'];
-  await runOasTest('TMF637-001-UnionTest.yaml', paths, 1, 4);
-});
-
-test('test_024_oas_test_014_testTMF637_TestRecursion', async () => {
-  const paths = [
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:#/c/s/Entity>prop:scalar:id',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:scalar:terminationDate',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:#/c/s/Extensible>prop:scalar:@baseType',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:#/c/s/Extensible>prop:scalar:@schemaLocation',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:#/c/s/Extensible>prop:scalar:@type',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:[inline:#/c/s/RelatedPartyOrPartyRole]>prop:scalar:role',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:[inline:#/c/s/RelatedPartyOrPartyRole]>prop:comp:partyOrPartyRole>union:#/c/s/PartyOrPartyRole>comp:type:#/c/s/Producer>comp:type:#/c/s/PartyRole>obj:type:#/c/s/Entity>prop:scalar:href',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:[inline:#/c/s/RelatedPartyOrPartyRole]>prop:comp:partyOrPartyRole>union:#/c/s/PartyOrPartyRole>comp:type:#/c/s/Producer>comp:type:#/c/s/PartyRole>obj:type:#/c/s/Entity>prop:scalar:id',
-    'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:[inline:#/c/s/RelatedPartyOrPartyRole]>prop:comp:partyOrPartyRole>union:#/c/s/PartyOrPartyRole>comp:type:#/c/s/Producer>comp:type:#/c/s/PartyRole>obj:type:[inline:#/c/s/PartyRole]>prop:scalar:name',
-    // 'get:/productById>res:r>comp:type:#/c/s/Product>obj:type:[inline:#/c/s/Product]>prop:array:#relatedParty>comp:type:#/c/s/RelatedPartyOrPartyRole>obj:type:[inline:#/c/s/RelatedPartyOrPartyRole]>prop:comp:partyOrPartyRole>union:#/c/s/PartyOrPartyRole>comp:type:#/c/s/Producer>comp:type:#/c/s/PartyRole>obj:type:[inline:#/c/s/PartyRole]>prop:circular-ref:#relatedParty'
-  ];
-
-  // expect.assertions(6);
-  const error = await runOasTest('TMF637-002-RecursionTest.yaml', paths, 1, 4);
-  // expect(error).toContain("Circular reference detected in `@connect(selection:)` on `Query.productById`");
-});
-
-test('test_048_oas_test_031_post-body-oneOf', async () => {
-  const paths = ['post:/event>**'];
-  await runOasTest(`post-sample.yaml`, paths, 3, 4);
-});
-
- */
-
-/*test('test_053_oas_test_036_time-series', async () => {
-  const paths = ['post:/market-data-services/time-series/search>**'];
-  await runOasTest('time-series-1.0.28.yaml', paths, 1, 12);
-});*/
-
-test('test_053_oas_test_036_time-series', async () => {
-  const paths = [
-    "post:/market-data-services/time-series/search>body:b>obj:input:#/c/s/PriceTimeSeriesRequestBody>prop:scalar:responseDataFormat",
-    "post:/market-data-services/time-series/search>res:r>obj:type:#/c/s/ResultCollection>prop:obj:_embedded>obj:type:_embedded>prop:array:#timeSeries>obj:type:#/c/s/PriceTimeSeries>prop:comp:dataPoint>union:dataPointUnion>obj:type:#/c/s/DataPointHighcharts>prop:scalar:dataFormat",
-    "post:/market-data-services/time-series/search>res:r>obj:type:#/c/s/ResultCollection>prop:obj:_embedded>obj:type:_embedded>prop:array:#timeSeries>obj:type:#/c/s/PriceTimeSeries>prop:comp:dataPoint>union:dataPointUnion>obj:type:#/c/s/DataPointHighcharts>prop:array:#dataPoints>array:Float"
-  ]
-  await runOasTest('time-series-1.0.28.yaml', paths, 1, 7);
-});
-
-/*
-test('test_028_oas_test_017_testMostPopularProduct_star', async () => {
-  const paths = ['get:/emailed/{period}.json>res:r>obj:type:emailedByPeriodJsonResponse>*'];
-  await runOasTest('most-popular-product.yaml', paths, 4, 1);
-});
 */
-
-
-/*
-test('test_054_oas_test-better-naming', async () => {
-  const paths = [
-    'get:/2.3.0/astronauts/>res:r>obj:type:#/c/s/PaginatedPolymorphicAstronautEndpointList>prop:scalar:count',
-    'get:/2.3.0/astronauts/>res:r>obj:type:#/c/s/PaginatedPolymorphicAstronautEndpointList>prop:array:#results>union:#/c/s/PolymorphicAstronautEndpoint>obj:type:#/c/s/AstronautDetailed>prop:comp:agency>comp:type:#/c/s/AgencyMini>obj:type:#/c/s/AgencyMini>prop:scalar:name',
-    'get:/2.3.0/astronauts/>res:r>obj:type:#/c/s/PaginatedPolymorphicAstronautEndpointList>prop:array:#results>union:#/c/s/PolymorphicAstronautEndpoint>obj:type:#/c/s/AstronautEndpointNormal>prop:comp:agency>comp:type:#/c/s/AgencyMini>obj:type:#/c/s/AgencyMini>prop:scalar:name'
-  ]
-
-  await runOasTest('launch_Library_2-docs-v2.3.0.json', paths, 116, 5);
-});*/